Continuing experiments with constant args and template type args
